// generated/src/app/api-models/tax-reporting-entity.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class TaxReportingEntity {
  taxRepEntity: string;
  description: string;
  payerName1: string;
  payerName2: string;
  payerAddress: string;
  payerCity: string;
  payerState: string;
  payerZipCode: string;
  transControlCode: string;
  transName: string;
  transAddress: string;
  transCity: string;
  transState: string;
  transZipCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  payerCountry: string;
  transCountry: string;
  payerPhone: string;
}